+++ juntagrico/entity/share.py
@@ -1,10 +1,8 @@
-# encoding: utf-8
-
 from django.utils.translation import gettext as _
 
 from juntagrico.entity.billing import *
+from juntagrico.lifecycle.share import check_share_consistency
 from juntagrico.util.bills import *
-from juntagrico.config import Config
 
 
 class Share(Billable):
@@ -22,13 +20,11 @@ class Share(Billable):
     notes = models.TextField(
         _('Notizen'), max_length=1000, default='', blank=True)
 
-    @classmethod
-    def create(cls, sender, instance, created, **kwds):
-        if created and Config.billing():
-            bill_share(instance)
+    def clean(self):
+        check_share_consistency(self)
 
     def __str__(self):
-        return _('Anteilschein #%s') % self.id
+        return _('Anteilschein {0}').format(self.id)
 
     class Meta:
         verbose_name = Config.vocabulary('share')