+++ juntagrico/templatetags/depot_extras.py
@@ -4,8 +4,8 @@ register = template.Library()
 
 
 @register.filter
-def subscriptions_amount(subscription, name):
-    return subscription.subscription_amount(name)
+def subscriptions_amount(subscription, id):
+    return subscription.subscription_amount(id)
 
 
 @register.filter
@@ -16,8 +16,3 @@ def extra_subscription(subscription, code):
 @register.filter
 def extra_subscription_amount(subscription, code):
     return subscription.extra_subscription_amount(code)
-
-
-@register.filter
-def get_item(dictionary, key):
-    return dictionary.get(key)