+++ test/test_cs.py
@@ -1,4 +1,5 @@
 from django.contrib import auth
+from django.urls import reverse
 
 from test.util.test import JuntagricoTestCase
 
@@ -9,34 +10,40 @@ class CreateSubscriptionTests(JuntagricoTestCase):
         self.client.force_login(self.member.user)
         user = auth.get_user(self.client)
         assert user.is_authenticated
-        self.client.get('/my/signup')
+        self.client.get('/my/signup/')
         self.assertEqual(str(auth.get_user(self.client)), 'AnonymousUser')
 
     def testRedirect(self):
-        response = self.client.get('/my/create/subscrition')
+        response = self.client.get(reverse('cs-subscription'))
         self.assertEqual(response.status_code, 302)
-        response = self.client.post('/my/create/subscrition')
-        self.assertEqual(response.status_code, 302)
-
-        response = self.client.get('/my/create/subscription/selectdepot')
-        self.assertEqual(response.status_code, 302)
-        response = self.client.post('/my/create/subscription/selectdepot')
+        response = self.client.post(reverse('cs-subscription'))
         self.assertEqual(response.status_code, 302)
 
-        response = self.client.get('/my/create/subscription/start')
+        response = self.client.get(reverse('cs-depot'))
         self.assertEqual(response.status_code, 302)
-        response = self.client.post('/my/create/subscription/start')
-        self.assertEqual(response.status_code, 302)
-
-        response = self.client.get('/my/create/subscription/addmembers')
-        self.assertEqual(response.status_code, 302)
-        response = self.client.post('/my/create/subscription/addmembers')
+        response = self.client.post(reverse('cs-depot'))
         self.assertEqual(response.status_code, 302)
 
-        response = self.client.get('/my/create/subscription/shares')
+        response = self.client.get(reverse('cs-start'))
         self.assertEqual(response.status_code, 302)
-        response = self.client.put('/my/create/subscription/shares')
+        response = self.client.post(reverse('cs-start'))
         self.assertEqual(response.status_code, 302)
 
-        response = self.client.get('/my/welcome')
+        response = self.client.get(reverse('cs-co-members'))
+        self.assertEqual(response.status_code, 302)
+        response = self.client.post(reverse('cs-co-members'))
+        self.assertEqual(response.status_code, 302)
+
+        response = self.client.get(reverse('cs-shares'))
+        self.assertEqual(response.status_code, 302)
+        response = self.client.put(reverse('cs-shares'))
+        self.assertEqual(response.status_code, 302)
+
+        response = self.client.get(reverse('cs-summary'))
+        self.assertEqual(response.status_code, 302)
+        response = self.client.put(reverse('cs-summary'))
+        self.assertEqual(response.status_code, 302)
+
+    def testWelcome(self):
+        response = self.client.get(reverse('welcome'))
         self.assertEqual(response.status_code, 302)