+++ test/test_jobs.py
@@ -1,34 +1,36 @@
+from django.urls import reverse
+
 from test.util.test import JuntagricoTestCase
 
 
 class JobTests(JuntagricoTestCase):
 
     def testAssignments(self):
-        self.assertSimpleGet('/my/assignments')
+        self.assertSimpleGet(reverse('jobs'))
 
     def testAssignmentsAll(self):
-        self.assertSimpleGet('/my/assignments/all')
+        self.assertSimpleGet(reverse('jobs-all'))
 
     def testJob(self):
-        self.assertSimpleGet('/my/jobs/' + str(self.job1.pk) + '/')
+        self.assertSimpleGet(reverse('job', args=[self.job1.pk]))
 
     def testPastJob(self):
-        self.assertSimpleGet('/my/pastjobs')
+        self.assertSimpleGet(reverse('memberjobs'))
 
     def testParticipation(self):
-        self.assertSimpleGet('/my/participation')
+        self.assertSimpleGet(reverse('areas'))
 
     def testTeam(self):
-        self.assertSimpleGet('/my/area/' + str(self.area.pk) + '/')
+        self.assertSimpleGet(reverse('area', args=[self.area.pk]))
 
     def testAreaJoinAndLeave(self):
-        self.assertSimpleGet('/my/area/' + str(self.area.pk) + '/join')
+        self.assertSimpleGet(reverse('area-join', args=[self.area.pk]))
         self.assertEqual(self.area.members.count(), 1)
-        self.assertSimpleGet('/my/area/' + str(self.area.pk) + '/leave')
+        self.assertSimpleGet(reverse('area-leave', args=[self.area.pk]))
         self.assertEqual(self.area.members.count(), 0)
 
     def testJobPost(self):
         self.client.force_login(self.member.user)
-        response = self.client.post('/my/jobs/' + str(self.job1.pk) + '/', {'jobs': 1})
+        response = self.client.post(reverse('job', args=[self.job1.pk]), {'jobs': 1})
         self.assertEqual(response.status_code, 302)
         self.assertEqual(self.job1.free_slots(), 0)