+++ test/test_subs.py
@@ -1,3 +1,5 @@
+from django.urls import reverse
+
 from juntagrico.entity.share import Share
 from test.util.test import JuntagricoTestCase
 
@@ -5,26 +7,27 @@ from test.util.test import JuntagricoTestCase
 class JobTests(JuntagricoTestCase):
 
     def testSub(self):
-        self.assertSimpleGet('/my/subscription/detail/')
-        self.assertSimpleGet('/my/subscription/detail/' + str(self.sub.pk) + '/')
+        self.assertSimpleGet(reverse('sub-detail'))
+        self.assertSimpleGet(reverse('sub-detail-id', args=[self.sub.pk]))
 
     def testSubChange(self):
-        self.assertSimpleGet('/my/subscription/change/overview/' + str(self.sub.pk) + '/')
+        self.assertSimpleGet(reverse('sub-change', args=[self.sub.pk]))
 
     def testDepotChange(self):
-        self.assertSimpleGet('/my/subscription/change/depot/' + str(self.sub.pk) + '/')
-        self.assertSimplePost('/my/subscription/change/depot/' + str(self.sub.pk) + '/', {'depot': self.depot2.pk})
+        self.assertSimpleGet(reverse('depot-change', args=[self.sub.pk]))
+        self.assertSimplePost(reverse('depot-change', args=[self.sub.pk]), {'depot': self.depot2.pk})
         self.sub.refresh_from_db()
         self.assertEqual(self.sub.future_depot, self.depot2)
 
     def testSizeChange(self):
-        self.assertSimpleGet('/my/subscription/change/size/' + str(self.sub.pk) + '/')
-        self.assertSimplePost('/my/subscription/change/size/' + str(self.sub.pk) + '/', {'subscription': self.sub_type2.pk})
-        self.sub.refresh_from_db()
-        self.assertEqual(self.sub.future_types.all()[0], self.sub_type)
-        self.assertEqual(self.sub.future_types.count(), 1)
-        Share.objects.create(**self.share_data)
-        self.assertSimplePost('/my/subscription/change/size/' + str(self.sub.pk) + '/', {'subscription': self.sub_type2.pk})
-        self.sub.refresh_from_db()
-        self.assertEqual(self.sub.future_types.all()[0], self.sub_type2)
-        self.assertEqual(self.sub.future_types.count(), 1)
+        with self.settings(BUSINESS_YEAR_CANCELATION_MONTH=12):
+            self.assertSimpleGet(reverse('size-change', args=[self.sub.pk]))
+            self.assertSimplePost(reverse('size-change', args=[self.sub.pk]), {'amount[' + str(self.sub_type2.pk) + ']': 1})
+            self.sub.refresh_from_db()
+            self.assertEqual(self.sub.future_types.all()[0], self.sub_type)
+            self.assertEqual(self.sub.future_types.count(), 1)
+            Share.objects.create(**self.share_data)
+            self.assertSimplePost(reverse('size-change', args=[self.sub.pk]), {'amount[' + str(self.sub_type2.pk) + ']': 1})
+            self.sub.refresh_from_db()
+            self.assertEqual(self.sub.future_types.all()[0], self.sub_type2)
+            self.assertEqual(self.sub.future_types.count(), 1)